+++ subprojects/kotlinx-serialization/src/commonMain/kotlin/org/kotools/types/kotlinx/serialization/internal/SerialName.kt
@@ -1,6 +1,0 @@
-package org.kotools.types.kotlinx.serialization.internal
-
-import org.kotools.types.internal.Warning
-
-@Suppress(Warning.UNUSED)
-internal expect inline fun <reified T : Any> serialNameOf(): String