+++ subprojects/kotlinx-serialization/src/jsMain/kotlin/org/kotools/types/kotlinx/serialization/internal/JsSerialName.kt
@@ -1,6 +1,0 @@
-package org.kotools.types.kotlinx.serialization.internal
-
-import kotools.types.internal.simpleNameOf
-
-internal actual inline fun <reified T : Any> serialNameOf(): String =
-    simpleNameOf<T>()